backend/requirements.txt
astroid==2.12.13
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
defusedxml==0.7.1
dill==0.3.6
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
future==0.18.2
greenlet==2.0.1
idna==3.4
isort==5.11.2
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==2.6.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.3.1
PyJWT==2.6.0
pylint==2.15.8
python-dotenv==0.21.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.3.0
SQLAlchemy==1.4.45
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
